VAST.AI/requirements.txt
xdg
vastai
mysql-connector-python